backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.26.2
python-jose[cryptography]==3.3.0 